requirements/base.txt
# ============================================================
#  LMS Platform — Base Requirements
#  Django 5.0+ | Python 3.11+
# ============================================================

# Core Framework
Django==5.0.6
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Cache & Queue
redis==5.0.7
django-redis==5.4.0
celery==5.4.0
django-celery-beat==2.6.0
celery[redis]==5.4.0

# Real-time (WebSocket)
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Authentication
django-allauth==0.63.6
djangorestframework-simplejwt==5.3.1
django-otp==1.5.1

# Storage (DigitalOcean Spaces / S3)
django-storages==1.14.4
boto3==1.34.144

# REST API
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
django-cors-headers==4.4.0

# Forms & Templates
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-widget-tweaks==1.5.0

# Image Processing
Pillow==10.4.0

# Production Server
gunicorn==22.0.0
whitenoise==6.7.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
stripe==10.4.0

# GitHub Integration
PyGithub==2.3.0

# YouTube API
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Email
django-anymail[sendgrid]==11.1

# Celery Utilities
kombu==5.3.7

# Type Hints
typing-extensions==4.12.2
